requirements.txt
flask
scikit-learn
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorflow-cpu # Use this if you don’t have a GPU
